feature(authAppleId): auth manager
